Include requests library
I so didn't want to do this.
It seems silly to import 2+mb of code just to make a few http requests. But getting a proxy handler set up with urllib.request ended up being more tricky than using urllib2 (urllib.request doesn't create a 'global' opener in the way urllib2 does. Which is to say it isn't global at all).

So here we are. I'm not happy about it, but certain things ended up being much more streamlined this way, so there are a few more benefits that just functionality.