Add ability to import from Kodi and Plex
Kodi is able to pull data directly from the server

Plex requires a csv export since it only gives IMDB or TMDB id #s with an individual metadata api request. I think it is rather silly to fire off a few hundred api requests just to get one extra data field. The csv is a minor inconvenience, but I'm ok with that.