More flexible split of rTorrent body and headers
Still working on #20. I'm not entirely sure what is going on, but the user was unable to split the headers and body using the former regex expression. This will split on double new-lines and uses the last index as the response body. This way the only requirement is that the body is at the end of the response (if it isn't we have bigger problems) and that if there are headers they must be separated with a blank line.