No longer need to restart to apply changed to scheduler
This is the worst commit I've done yet.

When writing config, runs a check that determines if a scheduled task needs to be reloaded (ie change in frequency, time of day, etc). This then executes the reload() function for the task object.

But in order to do that imports had to be rearranged for reasons I don't really remember. So I decided to start moving some of the bulky ajax methods to a module where they make more sense. More of this needs to be done.

Long story short, a whole lot of stuff was refactored to make everything better and I don't feel like retracing my steps to document all of it.

So there.