Fix changing cwd *again*
I don't even know how this happened. After daemonizing in *nix the cwd would change to /, which is not cool so I'd change it back to the script's dir.

Except that something happened where it doesn't seem to stick and gets set to '/' for the life of the program. This fixes it.

If you have trouble with this update you may need to run Watcher *not* in daemon mode, update to this, then start again as usual.