Fix TPB sometimes not getting release title
I came across a result that wouldn't get the text contents of the <a> tag while parsing, even thoguh valid text was present. This change tries to get the anchor text, then the title attribute of the anchor tag. If neither are present it ignores the result and continues the loop.