Rewrite of scheduler plugin
Eliminate subclass Task -- it isn't neccesary and only makes it less readable.

Create pesistence record file. This is a json formatted file that contains keys for each task, the value of which is a dict. The nested dict will have a key 'lastexecution' which is a timestamp of the last time this task had run. This will be helpful in the future when long-interval tasks are scheduled. For example, a task that runs every other day would currently never run if the user updates every day and restarts the server. This update allows the scheduler to read the last executed date and calculate the next iteration off that time + intervals until it is the next possible execution.

Checks if task timer is alive before attempting to stop it. Doesn't affect performance, but gives better logging/stdout feedback.

More clear variable names

Better docs throughout