More scheduler plugins refinements
Contrary to what I was taught, file handlers do not seem to be thread safe and end up causing malformed JSON when simultansouly dumping to the file. A shared thread lock solves this.

Added a distinct write_record function.

If a record for a task doesn't exist, write a record for it for now. This way long-interval tasks won't be perpetually delayed if the server is updated/restarted before the first execution.