Add option for symlinks when post-processing
Post-processing settings now inlcudes a 'Move Method' option, which can be either Move, Hardlink, or Symbolic Link.

Move simply moves the movie.

Hardlink creates a hardlink in the directory provided.

Symlink moves the file to the destination directory then makes a symlink back to the original download.

Symlinks are only enabled for *nix users. Windows requires certain elevated privileges to create symlinks, so this setting must be enabled by manually editing the config (or un-disabling it in the DOM). Just a bit of weak gatekeeping.