Improvements to automatic metadata updater
Don't update metadata for Finished movies. If they are finished we can assume we have all of the information neccexary to get us to that point.

Add a logging line relaying movie titles that are in metadata update queue.

Only download poster if we don't have it. No need to re-download posters, it just adds more overhead to both us and TMDB. Posters are still forcibly redownloaded through the metadata update in the movie's status modal.