Fix posters being removed during metadata update
Missing posters will be automatially redownloaded during the next scheduled metadata update.

Fixes issues where posters were always removed, regardless of the method re-downloading it. Now posters are only removed if posters are to be redownloaded immediately afterward.

Add check for posters existing on disk. If a poster is supposed to be in /static/images/posters/ and isn't, will trigger a redownload.