When downloading, get hash if unknown for selected torrent
Torznab search results will sometimes not provide a hash for the torrent. In order to ID the torrent, the download link is used for the guid in place of the missing hash. This gets the hash for the torrent when the download is triggered so post-processing can be completed using the hash to identify the release.