Improvements to PreDB flow
Predb checking no longer sets the movie's status, this is handled by the searcher.verify method. This leaves the predb methods responsible only for updading predb fields in the database.

When searching for movies, pull from database after predb checks so verify can be called for each checked movie since it is now responible for bumping the status of the movie.