Add bitcoin donate address
So I've been mostly opposed to accepting donations until now. Even so I'm not thrilled about it, but this project takes up a fair amount of time and if someone wants to throw a dollar my direction who am I to stop them?

While I still suggest that users consider donating to a good charity before sending me anything, this allows the option.