#174 Switch to yahoo format with xmljson
Was previously using gdata format, but this would attempt to parse all contents as floats or ints before resorting to string. This then requires explicit string coercion when parsing.

Now using yahoo format all contents are strings. This requires explicit int coercion, which I feel is easier to read and manage. Might be several microseconds faster as well, but too little to be measurable.

This fixes the cause of the issue, which is a release with a numerical title being parsed as int. Now all data is a string by default and fewer checks need to be made.