Add Trakt RSS sync
Can sync user-supplied rss feed from Trakt (thanks @barbequesauce !)

This is not without its challenges. The time format in the rss feed (specifically timezone and offset sections) aseems to change almost randomly. So dates are parsed without those segments. I'm pretty sure this won't cause problems.

Syncs with the same frequency as the standard Trakt lists.

The feed doesn't provide an imdb or tmdb id #, so it is searched using the title and year. Under fringe cases this may cause the wrong movie to be added, but this would be incredibly rare and I'm not sure there is anything I can do about it.