Fix dict.get() default values
For some reason I always thoght that dict.get() would default to the second passed value if the dict's value for that key is None. But it only defaults to the second value if that key isn't present at all in the dict.

This fixes most (hopefully all) of those instances.

Minor cleanup and readability improvements throughout.