Fix for git and case-sensitive file names
This commit and the following commit are to correct file names for 3010f3b24fd085bde38d49532dd4cce68da5ca56

Git doesn't seem to recognize changes in file names if the only change is case. This pair of commits aims to address this by deleting the files, then recreating them with the correct name.