Merge origin/bootstrap4
Alright, a lot going on here. Bear with me.

Convert frontend from Bootstrap3 to Bootstrap4. This includes new themes (since BS4 uses scss instead of less), a lot of new javascript, a lot less css, and a much cleaner looking UI.

With javascript, I converted a lot to vanilla JS instead of using JQuery. This is done mostly for speed, but can also make the code a bit easier to read.

Much less CSS. Some pages don't need any specific styles.

Moved suggestions to the Add Movie page so all methods to add movies are in a tabbed interface.

Addresses #136. Any quality profile can be marked as the 'default' profile to be used when a profile is not specified. 

Addresses #190, #191. In library/manage a movie's database entry can be edited (to a degree). This can break stuff, so be careful.

Cleanup of code throughout. Too many little changes to list.

core.sql.update_multiple_values now uses an arbitrary idcol and idval for identifying rows to update.

I'm sure there are things I'm missing, but all of these fixes were easier to take care of at once.