#231 Add arbitrary filter params to liststatus API calls
Any filters can be applied to liststatus by passing &column=value using column names from the databse.

Doing this prompted a quick rewrite of the api dispatching method. It checks if there is a method with the same name as the mode param and sends all params except mode and api key to this method as dict in the first arg.