Basic framework
All commands are simple POST requests, so no need for an external library. The only abnormal parts are...

Getting the oauth token. I don't believe there is a code-only approach, so this must be done by the user in a browser.

Post-processing. They offer a callback, but I'm not sure exactly what info is passed in their POST to the callback url. Should be able to set aside a special api or postprocessing method to handle this incoming data.