Where my classes gone?
This is a larger commit than I really wanted, but all of this kind of had to be done at once.

For whatever reason, when creating different modules I put classes everywhere. This requires instances of all of these classes in order to access methods that had no need to reference the self object. This commit removes all classes that are possible to remove and makes use of static methods for when several namespaces are useful in a single module.

As is typical, several other adjustments were made along the way.

searchresults.py :
Exit score() call immediately if there are no releases to score rather than calling the individual scoring methods on an empty list.

postprocessing.py :
Fixed a bug where torrent hashes wouldn't match if not all lowercase.

library.py / movieinfo.py :
Move Poster methods to movieinfo as Poster.save and Poster.remove

library.py :
Fix reading movie status when import releases contain the latest possible status value.

api.py :
Return stringified json when api calls return an error from the default method.

To anyone who actually takes the time to read this... I apologize for the huge, messy commit. It had to be done.