'#260 & #247 stop messing up files
Using a hash function with lower collision over the whole file name prevents from confusing between a new file and an older one that would have had the same hash.

Current hash was computed upon the first 16 chars of the file name so if ever there is a common prefix (such as "[myfevoriteindexer.com]")  the hash  is the same.

Now with the full file name being used it reduce the risk of collision which will avoid to retrieve old data from previous run (#247)

Furthermose having the same guid convention in postprocessing (lower case) it enables to find back a previous processing and avoid seeking again for movie metadata when files are not deleted from download directory (#260)