#257 Fix parsing errors with specific movie titles
When a movie's title contained a substring such as 'cam' in 'Camelot' it would get parsed as CAM quality regardless of the actual quality. This commit correct this by splitting the release name on the movie's year and parsing only the right half of this split for quality information.

This also calls lower() on this data at the beginning and matches all terms in lower-case rather than calling upper() for resolution and lower() for source.