#258 Add task to scan for missing files
Adds scheduledtask that scans disk for movie files and applies an action if they are not found. Currently two options are possible: revert or remove.

Revert is similar to a reset done through the management page except that it does not reset predb flags.

Remove removes movies from the library if their file has been deleted.